basic socket

+++ src/main/java/org/example/java_lab/Point.java
@@ -1,26 +1,26 @@
 package org.example.java_lab;
 
 public class Point {
-    private float x = 0, y = 0;
+    private double x = 0, y = 0;
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
